Update to reflect new flash dumper primitive.
Only on this branch at the moment.

publish-aa- is now quite a few commits behind everything else.

Local branches (unpublished) are more recent than publish-aa- is,
but not as recent as this (present) branch is.

master is reasonably recent, as it already has a sample flash
recite routine to start it (not really a dump in the vein of
say, 'hexdump -C' which is what the current dump word does
with SPI flashROM).

On branch lark-ag-fork-FLASHDUMP-a-

	modified:   README.md